requirements.txt
fastapi==0.118.2
matplotlib==3.10.3
numpy==1.26.4
pillow==11.3.0
prometheus_client==0.22.1
pydantic==2.11.7
python-jose==3.5.0
starlette==0.48.0
torchaudio==2.7.1
torch==2.4.1
torchvision==0.22.0
torchvision-extra-decoders==0.0.2
uvicorn==0.35.0
